requirements.txt
# ============================================================
# requirements.txt — Fetal Image Analysis Inference Package
# ============================================================
# Install: pip install -r requirements.txt
# ============================================================

ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
flask>=3.0.0
flask-cors>=4.0.0
pyyaml>=6.0
tqdm>=4.65.0
